requirements_airflow.txt
# Apache Airflow requirements for EduSight
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-email==1.4.0
apache-airflow-providers-http==4.9.0
apache-airflow-providers-celery==3.6.0
apache-airflow-providers-redis==3.5.0

# Additional providers for EduSight integrations
apache-airflow-providers-mysql==5.5.0
apache-airflow-providers-sqlite==3.7.0
celery==5.3.4
redis==5.0.1

# Data processing
pandas==2.1.4
numpy==1.24.3

# For custom operators
requests==2.31.0
psutil==5.9.6
